Implemented App.java as a Main class

+++ JavaDeveloperTest-AymenNaili/src/main/java/de/goeuro/internship/model/Location.java
@@ -150,31 +150,31 @@ public class Location {
 	}
 
 	public class GeoPosition {
-		private int latitude;
-		private int longitude;
+		private double latitude;
+		private double longitude;
 
 		public GeoPosition() {
 		}
 
-		public GeoPosition(int latitude, int longitude) {
+		public GeoPosition(double latitude, double longitude) {
 			super();
 			this.latitude = latitude;
 			this.longitude = longitude;
 		}
 
-		public int getLatitude() {
+		public double getLatitude() {
 			return latitude;
 		}
 
-		public void setLatitude(int latitude) {
+		public void setLatitude(double latitude) {
 			this.latitude = latitude;
 		}
 
-		public int getLongitude() {
+		public double getLongitude() {
 			return longitude;
 		}
 
-		public void setLongitude(int longitude) {
+		public void setLongitude(double longitude) {
 			this.longitude = longitude;
 		}
 	}
